Changed the argument order of STRING-FILTER and STRING-DELETE, to conform to the reference implementation.
This is a big exception to our usual procedure, and only warranted by
a) the widespread adoption of the reference implementation and its
argument order b) an elaborate explanation on Olin why it's clear the
reference implementation is right and the document isn't.

This takes care of:

http://srfi.schemers.org/srfi-13/post-mail-archive/msg00002.html

posted by David Van Horn.